ML/requirements.txt
flask==3.0.3
flask-cors==5.0.0
scikit-learn==1.5.2
numpy==2.0.1
pandas==2.2.2
joblib==1.4.2
pymongo>=4.6.0
python-dotenv>=1.0.1
